# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MENTALMATE API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MoodLogMood = (typeof MoodLogMood)[keyof typeof MoodLogMood];

export const MoodLogMood = {
  happy: "happy",
  sad: "sad",
  stressed: "stressed",
  anxious: "anxious",
  calm: "calm",
  tired: "tired",
} as const;

export interface MoodLog {
  id: number;
  userId: string;
  mood: MoodLogMood;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  stressLevel?: number | null;
  date: string;
  createdAt: string;
}

export type CreateMoodLogMood =
  (typeof CreateMoodLogMood)[keyof typeof CreateMoodLogMood];

export const CreateMoodLogMood = {
  happy: "happy",
  sad: "sad",
  stressed: "stressed",
  anxious: "anxious",
  calm: "calm",
  tired: "tired",
} as const;

export interface CreateMoodLog {
  mood: CreateMoodLogMood;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  stressLevel?: number | null;
}

export interface DayMoodData {
  date: string;
  day: string;
  /** @nullable */
  mood?: string | null;
  /** @nullable */
  stressLevel?: number | null;
}

export interface MoodCounts {
  happy: number;
  sad: number;
  stressed: number;
  anxious: number;
  calm: number;
  tired: number;
}

export interface WeeklyMoodSummary {
  days: DayMoodData[];
  dominantMood: string;
  /** @nullable */
  averageStress?: number | null;
  insight: string;
  moodCounts: MoodCounts;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  userId: string;
  role: ChatMessageRole;
  content: string;
  isCrisis: boolean;
  createdAt: string;
}

export interface SendChatMessage {
  message: string;
}

export interface WellnessDashboard {
  userId: string;
  /** @nullable */
  stressLevel?: number | null;
  /** @nullable */
  bloodPressureSystolic?: number | null;
  /** @nullable */
  bloodPressureDiastolic?: number | null;
  /** @nullable */
  mentalHealthScore?: number | null;
  /** @nullable */
  lastUpdated?: string | null;
  streakDays: number;
  totalMoodLogs: number;
  totalChatSessions: number;
}

export interface UpdateWellnessData {
  /** @nullable */
  stressLevel?: number | null;
  /** @nullable */
  bloodPressureSystolic?: number | null;
  /** @nullable */
  bloodPressureDiastolic?: number | null;
  /** @nullable */
  mentalHealthScore?: number | null;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  mood_logged: "mood_logged",
  chat_session: "chat_session",
  wellness_updated: "wellness_updated",
  milestone: "milestone",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  createdAt: string;
}

export interface Doctor {
  id: number;
  name: string;
  specialization: string;
  rating: number;
  reviewCount: number;
  availability: string;
  location: string;
  phone: string;
  email: string;
  /** @nullable */
  imageUrl?: string | null;
  bio: string;
  acceptingPatients: boolean;
}

export type SymptomInputSeverity =
  (typeof SymptomInputSeverity)[keyof typeof SymptomInputSeverity];

export const SymptomInputSeverity = {
  mild: "mild",
  moderate: "moderate",
  severe: "severe",
} as const;

export interface SymptomInput {
  symptoms: string[];
  duration?: string;
  severity: SymptomInputSeverity;
}

export type PredictedConditionLikelihood =
  (typeof PredictedConditionLikelihood)[keyof typeof PredictedConditionLikelihood];

export const PredictedConditionLikelihood = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export interface PredictedCondition {
  name: string;
  likelihood: PredictedConditionLikelihood;
  description: string;
}

export interface ConditionPrediction {
  disclaimer: string;
  conditions: PredictedCondition[];
  recommendations: string[];
  seekHelpImmediately: boolean;
}