# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MENTALMATE API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mood:
    get:
      operationId: getMoodLogs
      tags: [mood]
      summary: Get mood logs for current user
      responses:
        "200":
          description: List of mood logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoodLog"
    post:
      operationId: createMoodLog
      tags: [mood]
      summary: Log today's mood
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMoodLog"
      responses:
        "201":
          description: Mood logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodLog"
  /mood/today:
    get:
      operationId: getTodayMood
      tags: [mood]
      summary: Get today's mood log
      responses:
        "200":
          description: Today's mood log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodLog"
        "404":
          description: No mood logged today
...
  /symptoms/predict:
    post:
      operationId: predictCondition
      tags: [symptoms]
      summary: Get possible condition based on symptoms (informational only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SymptomInput"
      responses:
        "200":
          description: Predicted condition suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionPrediction"
components:
  schemas:
...
    MoodLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        mood:
          type: string
          enum: [happy, sad, stressed, anxious, calm, tired]
        note:
          type: ["string", "null"]
        stressLevel:
          type: ["integer", "null"]
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - mood
        - date
        - createdAt
    CreateMoodLog:
      type: object
      properties:
        mood:
          type: string
          enum: [happy, sad, stressed, anxious, calm, tired]
        note:
          type: ["string", "null"]
        stressLevel:
          type: ["integer", "null"]
      required:
        - mood
...
    SymptomInput:
      type: object
      properties:
        symptoms:
          type: array
          items:
            type: string
        duration:
          type: string
        severity:
          type: string
          enum: [mild, moderate, severe]
      required:
        - symptoms
        - severity
    ConditionPrediction:
      type: object
      properties:
        disclaimer:
          type: string
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/PredictedCondition"
        recommendations:
          type: array
          items:
            type: string
        seekHelpImmediately:
          type: boolean
      required:
        - disclaimer
        - conditions
        - recommendations
        - seekHelpImmediately
    PredictedCondition:
      type: object
      properties:
        name:
          type: string
        likelihood:
          type: string
          enum: [low, moderate, high]
        description:
          type: string
      required:
        - name
        - likelihood
        - description